mm fixed pm fixed mm profile fixed done with project

+++ lib/employee/emp_create_order_page.dart
@@ -635,7 +635,7 @@ class _CattleFeedOrderScreenState extends State<CattleFeedOrderScreen> {
               ),
 
             const SizedBox(height: 40),
-            _buildOrderSummaryCard(totalWeight as int , unit, pricePerBag, totalPrice as int),
+            _buildOrderSummaryCard(totalWeight as int , unit, pricePerBag, totalPrice),
             const SizedBox(height: 30),
           ],
         ),
